# proto: amortize cost of growing a Buffer