Fixes https://github.com/golang/protobuf/issues/566