@rhysh 2018-04-06
The new `grow` method doubles the capacity, unless there's a need for more (or if there's no need at all). The `runtime.growslice` function (used for growth during `append`) often doubles, but will increase the size by only 25% if the slice starts off at 1024 or more elements. This means it won't hog too much memory but will still amortize the cost of the allocations/moves. Go's `append` has done that [since the beginning](https://github.com/golang/go/commit/d8b5d039cd1bec151cc325973ff32bd34ebb0456). Is that complexity worthwhile here, or is doubling acceptable?

Here's the result of the Go benchmarks from https://github.com/google/protobuf. `benchstat` reports no statistically significant changes, but a total geomean change of +0.21%.

```
name                                                                              old time/op  new time/op  delta
/./datasets/google_message1/proto2/dataset.google_message1_proto2.pb/Unmarshal-8  1.36µs ± 8%  1.38µs ±11%    ~     (p=0.529 n=10+10)
/./datasets/google_message1/proto2/dataset.google_message1_proto2.pb/Marshal-8    1.00µs ±10%  1.00µs ± 9%    ~     (p=0.895 n=10+10)
/./datasets/google_message1/proto2/dataset.google_message1_proto2.pb/Size-8        292ns ± 6%   297ns ± 9%    ~     (p=0.470 n=10+10)
/./datasets/google_message1/proto2/dataset.google_message1_proto2.pb/Clone-8      8.91µs ± 5%  8.97µs ± 6%    ~      (p=0.780 n=10+9)
/./datasets/google_message1/proto2/dataset.google_message1_proto2.pb/Merge-8      8.72µs ± 8%  8.68µs ± 7%    ~     (p=0.912 n=10+10)
/./datasets/google_message1/proto3/dataset.google_message1_proto3.pb/Unmarshal-8  1.00µs ±11%  0.98µs ± 8%    ~     (p=0.436 n=10+10)
/./datasets/google_message1/proto3/dataset.google_message1_proto3.pb/Marshal-8    1.33µs ± 7%  1.30µs ± 3%    ~     (p=0.343 n=10+10)
/./datasets/google_message1/proto3/dataset.google_message1_proto3.pb/Size-8        369ns ± 8%   368ns ± 2%    ~      (p=0.203 n=10+9)
/./datasets/google_message1/proto3/dataset.google_message1_proto3.pb/Clone-8      7.02µs ± 7%  6.97µs ± 5%    ~     (p=1.000 n=10+10)
/./datasets/google_message1/proto3/dataset.google_message1_proto3.pb/Merge-8      6.78µs ± 2%  6.84µs ± 7%    ~       (p=0.863 n=9+9)
/./datasets/google_message2/dataset.google_message2.pb/Unmarshal-8                 601µs ± 8%   606µs ± 7%    ~     (p=0.684 n=10+10)
/./datasets/google_message2/dataset.google_message2.pb/Marshal-8                   355µs ± 8%   353µs ± 4%    ~     (p=0.971 n=10+10)
/./datasets/google_message2/dataset.google_message2.pb/Size-8                     82.1µs ± 2%  82.8µs ± 2%    ~     (p=0.190 n=10+10)
/./datasets/google_message2/dataset.google_message2.pb/Clone-8                    4.20ms ± 3%  4.23ms ± 3%    ~       (p=0.370 n=8+9)
/./datasets/google_message2/dataset.google_message2.pb/Merge-8                    4.29ms ± 4%  4.36ms ± 5%    ~      (p=0.243 n=9+10)
[Geo mean]                                                                        14.4µs       14.4µs       +0.21%
```