@dsnet 2018-04-13
Thanks for filing the issue and PR.

The doubling behavior is probably fine. Fewer and fewer code paths go through the `Buffer` type, so I'm not too worried about that.