@rhysh 2018-04-17
Thanks for the test structure advice, @dsnet . Please take another look.

With the old growth behavior, the test shows a 500x ratio for allocating 3GB in order to add 1000 6-byte messages to a `proto.Buffer`:

```
--- FAIL: TestBufferMarshalAllocs (0.00s)
	all_test.go:401: needed 60B, allocated 330B total (ratio 5.5), allocated 330B at runtime (ratio 5.5)
	all_test.go:401: needed 120B, allocated 1260B total (ratio 10.5), allocated 1260B at runtime (ratio 10.5)
	all_test.go:401: needed 300B, allocated 7650B total (ratio 25.5), allocated 7650B at runtime (ratio 25.5)
	all_test.go:401: needed 600B, allocated 30300B total (ratio 50.5), allocated 30300B at runtime (ratio 50.5)
	all_test.go:401: needed 1200B, allocated 120600B total (ratio 100.5), allocated 120600B at runtime (ratio 100.5)
	all_test.go:401: needed 3000B, allocated 751500B total (ratio 250.5), allocated 751500B at runtime (ratio 250.5)
	all_test.go:401: needed 6000B, allocated 3003000B total (ratio 500.5), allocated 3003000B at runtime (ratio 500.5)
	all_test.go:401: needed 300B, allocated 6934B total (ratio 23.1), allocated 6834B at runtime (ratio 22.8)
	all_test.go:401: needed 600B, allocated 29584B total (ratio 49.3), allocated 29484B at runtime (ratio 49.1)
	all_test.go:401: needed 1200B, allocated 119884B total (ratio 99.9), allocated 119784B at runtime (ratio 99.8)
	all_test.go:401: needed 3000B, allocated 750784B total (ratio 250.3), allocated 750684B at runtime (ratio 250.2)
	all_test.go:401: needed 6000B, allocated 3002284B total (ratio 500.4), allocated 3002184B at runtime (ratio 500.4)
```